+++ tests/Stub/fixtures/phpdoc-fqcn/expected.php
@@ -5,7 +5,7 @@ namespace Test\CommandBuilder;
 interface CommandBuilderInterface
 {
     /**
-     * @return string|array<string|\Stringable|int>
+     * @return string|array<string|\Stringable|\Symfony\Component\Console\Command\Command|int>
      */
     public function getCommand(): array|string;
 }