+++ tests/Stub/fixtures/phpdoc-template/expected.php
@@ -2,13 +2,12 @@
 
 namespace Test\Foobar;
 
-use Castor\Container;
 /**
  * @template TKey of key-of<ContextData>
  * @template TDefault
  *
  * @param TKey|string $key
- * @param TDefault    $default
+ * @param TDefault $default
  *
  * @phpstan-return ($key is TKey ? ContextData[TKey] : TDefault)
  */