+++ tests/Stub/fixtures/use-alias/input1.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Test;
+
+use Castor\Helper\PathHelper;
+
+class Bar
+{
+    public function bar(PathHelper $pathHelper): void
+    {
+    }
+}